fix proto chain
