+++ src/mixins/object_ancestry.mixin.ts
@@ -5,8 +5,6 @@ import type { StaticCanvas } from '../__types__';
 
 type TAncestor = FabricObject | StaticCanvas;
 
-type TContainer = Group | StaticCanvas;
-
 /**
  * Strict: only ancestors that are objects (without canvas)
  */
@@ -157,42 +155,5 @@ export class FabricObjectAncestryMixin {
   ): boolean {
     const commonAncestors = this.findCommonAncestors(other, strict);
     return commonAncestors && !!commonAncestors.common.length;
-  }
-
-  /**
-   *
-   * @param {FabricObject} other object to compare against
-   * @returns {boolean | undefined} if objects do not share a common ancestor or they are strictly equal it is impossible to determine which is in front of the other; in such cases the function returns `undefined`
-   */
-  isInFrontOf(
-    this: FabricObject & this,
-    other: FabricObject & this
-  ): boolean | undefined {
-    if (this === other) {
-      return undefined;
-    }
-    const ancestorData = this.findCommonAncestors(other);
-    if (!ancestorData) {
-      return undefined;
-    }
-    if (ancestorData.fork.includes(other as any)) {
-      return true;
-    }
-    if (ancestorData.otherFork.includes(this as any)) {
-      return false;
-    }
-    const firstCommonAncestor = ancestorData.common[0];
-    if (!firstCommonAncestor) {
-      return undefined;
-    }
-    const headOfFork = ancestorData.fork.pop() as FabricObject,
-      headOfOtherFork = ancestorData.otherFork.pop() as FabricObject,
-      thisIndex = (firstCommonAncestor as TContainer)._objects.indexOf(
-        headOfFork
-      ),
-      otherIndex = (firstCommonAncestor as TContainer)._objects.indexOf(
-        headOfOtherFork
-      );
-    return thisIndex > -1 && thisIndex > otherIndex;
   }
 }
