+++ src/mixins/object_stacking.mixin.ts
@@ -1,9 +1,13 @@
 // @ts-nocheck
-
 import { fabric } from '../../HEADER';
 import type { FabricObject } from '../shapes/fabricObject.class';
+import type { Group } from '../shapes/group.class';
+import type { StaticCanvas } from '../__types__';
+import { FabricObjectAncestryMixin } from './object_ancestry.mixin';
 
-export class FabricObjectObjectStackingMixin {
+type TContainer = Group | StaticCanvas;
+
+export class FabricObjectObjectStackingMixin extends FabricObjectAncestryMixin {
   /**
    * Moves an object to the bottom of the stack of drawn objects
    * @return {FabricObject} thisArg
@@ -83,5 +87,42 @@ export class FabricObjectObjectStackingMixin {
       this.canvas.moveTo(this, index);
     }
     return this;
+  }
+
+  /**
+   *
+   * @param {FabricObject} other object to compare against
+   * @returns {boolean | undefined} if objects do not share a common ancestor or they are strictly equal it is impossible to determine which is in front of the other; in such cases the function returns `undefined`
+   */
+  isInFrontOf(
+    this: FabricObject & this,
+    other: FabricObject & this
+  ): boolean | undefined {
+    if (this === other) {
+      return undefined;
+    }
+    const ancestorData = this.findCommonAncestors(other);
+    if (!ancestorData) {
+      return undefined;
+    }
+    if (ancestorData.fork.includes(other as any)) {
+      return true;
+    }
+    if (ancestorData.otherFork.includes(this as any)) {
+      return false;
+    }
+    const firstCommonAncestor = ancestorData.common[0];
+    if (!firstCommonAncestor) {
+      return undefined;
+    }
+    const headOfFork = ancestorData.fork.pop() as FabricObject,
+      headOfOtherFork = ancestorData.otherFork.pop() as FabricObject,
+      thisIndex = (firstCommonAncestor as TContainer)._objects.indexOf(
+        headOfFork
+      ),
+      otherIndex = (firstCommonAncestor as TContainer)._objects.indexOf(
+        headOfOtherFork
+      );
+    return thisIndex > -1 && thisIndex > otherIndex;
   }
 }
