+++ src/shapes/fabricObject.class.ts
@@ -1,6 +1,5 @@
 import { fabric } from '../../HEADER';
 import { FabricObjectSVGExportMixin } from '../mixins/object.svg_export';
-import { FabricObjectAncestryMixin } from '../mixins/object_ancestry.mixin';
 import { FabricObjectObjectAnimationMixin } from '../mixins/object_animation.mixin';
 import { InteractiveFabricObject } from '../mixins/object_interactivity.mixin';
 import { FabricObjectObjectStackingMixin } from '../mixins/object_stacking.mixin';
@@ -10,7 +9,6 @@ import { applyMixins } from '../util/applyMixins';
 // TODO somehow we have to make a tree-shakeable import
 
 applyMixins(InteractiveFabricObject, [
-  FabricObjectAncestryMixin,
   FabricObjectObjectStackingMixin,
   FabricObjectObjectStraighteningMixin,
   FabricObjectSVGExportMixin,
